dom/media/webaudio/test/test_scriptProcessorNodeZeroInputOutput.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioBufferSourceNode</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var context = new AudioContext();

  var sp = context.createScriptProcessor(2048, 0, 2);
  sp.onaudioprocess = function(e) {
    is(e.inputBuffer.numberOfChannels, 0, "Should have 0 input channels");
    is(e.outputBuffer.numberOfChannels, 2, "Should have 2 output channels");
    sp.onaudioprocess = null;

    sp = context.createScriptProcessor(2048, 2, 0);
    sp.onaudioprocess = function(e) {
      is(e.inputBuffer.numberOfChannels, 2, "Should have 2 input channels");
      is(e.outputBuffer.numberOfChannels, 0, "Should have 0 output channels");
      sp.onaudioprocess = null;

      SimpleTest.finish();
    };
    sp.connect(context.destination);
  };
  sp.connect(context.destination);
});

</script>
</pre>
</body>
</html>
